+++ order/src/main/java/com/keepgoing/order/infrastructure/persistence/order/entity/Hub.java
@@ -1,13 +1,14 @@
-package com.keepgoing.order.infrastructure.persistence.entity.order;
+package com.keepgoing.order.infrastructure.persistence.order.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
+import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member {
-    @Column(name="member_id", nullable = false)
-    private Long memberId;
+public class Hub {
+    @Column(name = "hub_id")
+    private UUID hubId;
 }
