+++ order/src/main/java/com/keepgoing/order/infrastructure/persistence/order/entity/Member.java
@@ -1,14 +1,13 @@
-package com.keepgoing.order.infrastructure.persistence.entity.order;
+package com.keepgoing.order.infrastructure.persistence.order.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Vendor {
-    @Column(name = "vendor_id", nullable = false)
-    private UUID vendorId;
+public class Member {
+    @Column(name="member_id", nullable = false)
+    private Long memberId;
 }
